Refactor the v1-specific code to a subdirectory.
