+++ back-end/package.json
@@ -6,7 +6,7 @@
   "license": "AGPL v3",
   "type": "module",
   "engines": {
-    "node": "18.x"
+    "node": "20.x"
   },
   "main": "dist/server.js",
   "scripts": {
@@ -33,6 +33,6 @@
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
     "eslint": "^8.45.0",
-    "typescript": "^5.1.6"
+    "typescript": "^5.3.3"
   }
 }
