+++ back-end/src/server.ts
@@ -3,50 +3,17 @@
 // See the LICENSE file for details.
 
 import express from 'express'
-import cookieParser from 'cookie-parser'
-import cookieSession from 'cookie-session'
 
-import {
-    apnsReceivedNotification,
-    apnsToken,
-    pubSubTokenRequest,
-    subscribeTokenRequest,
-    subscribeToPublisher
-} from './routes.js';
-import {getDb, getSessionKeys} from './db.js'
-import {loadSettings} from './settings.js'
+import {v1router} from './v1/router.js'
+import {subscribeToPublisher} from './v1/routes.js';
+import {asyncWrapper, cookieMiddleware, sessionMiddleware} from './middleware.js'
 
 const PORT = process.env.PORT || 5001;
-
-loadSettings()
-await getDb()
-
-const cookieMiddleware = cookieParser()
-const sessionMiddleware = cookieSession({ keys: await getSessionKeys() })
 
 express()
     .use(express.json())
     .use(express.static('public'))
-    .post('/api/apnsToken', asyncWrapper(apnsToken))
-    .post('/api/apnsReceivedNotification', asyncWrapper(apnsReceivedNotification))
-    .post('/api/pubSubTokenRequest', asyncWrapper(pubSubTokenRequest))
+    .use('/api/v1', v1router)
+    .use('/api', v1router)
     .get('/subscribe/:publisherId', [cookieMiddleware, sessionMiddleware], asyncWrapper(subscribeToPublisher))
-    .get('/api/subscribeTokenRequest', [sessionMiddleware], asyncWrapper(subscribeTokenRequest))
     .listen(PORT, () => console.log(`Listening on port ${PORT}`))
-
-type Handler = (req: express.Request, res: express.Response, next: express.NextFunction) => Promise<void>
-
-function asyncWrapper(handler: Handler) {
-    return async (req: express.Request, res: express.Response, next: express.NextFunction) => {
-        try {
-            await handler(req, res, next)
-        }
-        catch (error) {
-            console.log(`Route handler produced an error: ${error}`)
-            if (!res.headersSent) {
-                // make sure we send some response to the client, if it's not already gone
-                res.status(500).send({ status: 'error', reason: `Server error: ${error}`})
-            }
-        }
-    }
-}
